refactor: DTO 레벨에서의 유효성 검사 추가, import 에러 수정
