+++ src/main/java/com/samsamhajo/deepground/studyGroup/dto/StudyGroupCommentRequest.java
@@ -16,5 +16,6 @@ public class StudyGroupCommentRequest {
   private Long studyGroupId;
 
   @NotBlank(message = "댓글 내용은 비어 있을 수 없습니다.")
+  @Size(max = 300, message = "댓글은 300자 이하로 입력해주세요.")
   private String content;
 }
