+++ src/main/java/com/samsamhajo/deepground/studyGroup/dto/StudyGroupCreateRequest.java
@@ -19,9 +19,11 @@ import lombok.NoArgsConstructor;
 public class StudyGroupCreateRequest {
 
   @NotBlank(message = "제목은 필수입니다.")
+  @Size(min = 2, max = 30, message = "제목은 2자 이상 30자 이하로 입력해주세요.")
   private String title;
 
   @NotBlank(message = "설명은 필수입니다.")
+  @Size(max = 500, message = "설명은 500자 이하로 입력해주세요.")
   private String explanation;
 
   @NotNull(message = "스터디 시작일은 필수입니다.")
@@ -37,8 +39,9 @@ public class StudyGroupCreateRequest {
   private LocalDate recruitEndDate;
 
   @NotNull(message = "정원은 필수입니다.")
-  @Min(value = 1, message = "정원은 최소 1명 이상이어야 합니다.")
-  private Integer groupMemberCount;
+  @Min(value = 2, message = "정원은 최소 2명 이상이어야 합니다.")
+  @Max(value = 10, message = "정원은 최대 10명까지 가능합니다.")
+  private Integer maxMembers;
 
   @NotNull(message = "오프라인 여부는 필수입니다.")
   private Boolean isOffline;
