+++ src/main/java/com/samsamhajo/deepground/studyGroup/exception/StudyGroupErrorCode.java
@@ -1,6 +1,7 @@
 package com.samsamhajo.deepground.studyGroup.exception;
 
-import com.samsamhajo.deepground.global.error.ErrorCode;
+
+import com.samsamhajo.deepground.global.error.core.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -12,4 +13,14 @@ public enum StudyGroupErrorCode implements ErrorCode {
 
     private final HttpStatus httpStatus;
     private final String message;
+
+    @Override
+    public HttpStatus getStatus() {
+        return httpStatus;
+    }
+
+    @Override
+    public String getMessage() {
+        return "[ERROR]" + message;
+    }
 }
