+++ src/main/java/com/samsamhajo/deepground/studyGroup/exception/StudyGroupException.java
@@ -1,9 +1,10 @@
 package com.samsamhajo.deepground.studyGroup.exception;
 
-import com.samsamhajo.deepground.global.error.ErrorCode;
-import com.samsamhajo.deepground.global.error.exception.BusinessException;
 
-public class StudyGroupException extends BusinessException {
+import com.samsamhajo.deepground.global.error.core.BaseException;
+import com.samsamhajo.deepground.global.error.core.ErrorCode;
+
+public class StudyGroupException extends BaseException {
     public StudyGroupException(ErrorCode errorCode) {
         super(errorCode);
     }
